# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warrant API
  description: REST API for Warrant ticketing (orgs, projects, tickets, reviews, queue).
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /projects/{projectID}:
    get:
      operationId: GetProject
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateProject
      summary: Update project (e.g. set status to active or closed)
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request (e.g. invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /tickets/{ticketID}:
    get:
      operationId: GetTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        "204":
          description: No content
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /projects/{projectID}/reviews:
    get:
      operationId: ListPendingReviews
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingReviewsResponse'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /tickets/{ticketID}/reviews:
    post:
      operationId: CreateReview
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /projects/{projectID}/escalations:
    get:
      operationId: ListEscalations
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Escalation'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /tickets/{ticketID}/escalations/{escalationID}/resolve:
    post:
      operationId: ResolveEscalation
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
        - name: escalationID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveEscalationRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /projects/{projectID}/queue/claim:
    post:
      operationId: ClaimTicket
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponseBody'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /tickets/{ticketID}/lease/renew:
    post:
      operationId: RenewLease
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewLeaseRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewLeaseResponseBody'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /tickets/{ticketID}/lease:
    delete:
      operationId: ReleaseLease
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
        - name: lease_token
          in: query
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lease_token:
                  type: string
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StructuredError:
      type: object
      required: [error, code, retriable]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [lease_expired, unauthorized, not_found, conflict, invalid_input, forbidden, internal, not_implemented, project_closed]
        retriable:
          type: boolean
...
    Project:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        repo_url:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        context_pack:
          type: object
        status:
          type: string
          enum: [active, closed]
          description: active (default) or closed; list endpoints default to active only.
        created_at:
          type: string
          format: date-time
    UpdateProjectRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
...
    Objective:
      type: object
      properties:
        description:
          type: string
        success_criteria:
          type: array
          items:
            type: string
        acceptance_test:
          type: string
    TicketContext:
      type: object
      properties:
        relevant_files:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        prior_attempts:
          type: array
          items:
            type: object
        human_answers:
          type: array
          items:
            type: string
    Ticket:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [task, bug, spike, review]
        priority:
          type: integer
          minimum: 0
          maximum: 3
        state:
          type: string
          enum: [pending, claimed, executing, awaiting_review, done, blocked, needs_human, failed]
        version:
          type: integer
        objective:
          $ref: '#/components/schemas/Objective'
        ticket_context:
          $ref: '#/components/schemas/TicketContext'
        inputs:
          type: object
        outputs:
          type: object
        depends_on:
          type: array
          items:
            type: string
        assigned_to:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateTicketRequest:
      type: object
      properties:
        depends_on:
          type: array
          items:
            type: string
    CreateReviewRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        notes:
          type: string
        reviewer_id:
          type: string
...
    PendingReviewsResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
    Escalation:
      type: object
      properties:
        id:
          type: string
        ticket_id:
          type: string
        agent_id:
          type: string
        reason:
          type: string
        question:
          type: string
        answer:
          type: string
        resolved_by:
          type: string
        resolved_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    ResolveEscalationRequest:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        reviewer_id:
          type: string
    ClaimRequest:
      type: object
      required: [agent_id]
      properties:
        agent_id:
          type: string
        priority:
          type: integer
          minimum: 0
          maximum: 3
        idempotency_key:
          type: string
          description: Optional. Retries with the same key return the same ticket/lease (renewed or re-claimed).
    ClaimResponseBody:
      type: object
      properties:
        ticket:
          $ref: '#/components/schemas/Ticket'
        lease:
          $ref: '#/components/schemas/Lease'
    Lease:
      type: object
      properties:
        ticket_id:
          type: string
        agent_id:
          type: string
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        renewable:
          type: boolean
    RenewLeaseRequest:
      type: object
      required: [lease_token]
      properties:
        lease_token:
          type: string
    RenewLeaseResponseBody:
      type: object
      properties:
        expires_at:
          type: string
          format: date-time